pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.5.0",
    "chromadb>=1.0.8",
    "fastapi>=0.115.9",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.16",
    "openai>=1.78.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
    "uvicorn>=0.34.2",
]
